resume_analyzer_backend/requirements.txt
Flask
Flask-CORS
pdfminer.six
python-docx
openai
